platform: the story of a historic mess
The platform information should be used to identify a platform
type, not the product name.  Not sure why pfSense came up with
this, but now it's so deeply tied into the system that we can't
just take it out.  The platform type is also steered by the
installer or build tools at the moment.  We will change that,
but it'll take more testing.

Discussed with: Jos