vpn: is_array($x) && is_array($x) -> isset($x)
I guess this is a typo.  It makes sense to preceede is_array() with
isset() here since PHP throws a warning on unset variables nowadays.
However, is_array() should not be used if the following operation is
`foreach' as it would mask the fact that the array that is set is not
an array at all (maybe due to a bug).  Rather want to catch it instead.

A bit of style along the way and `lint' target fixing.