completely tool-based approach with xgettext(1)
xgettext(1) is obviously much simpler than creating the canonical .pot file by hand.  (If only I'd have known better before I started this :)

This .pot file was generated from the source code via:

  # for i in `git grep --name-only 'gettext' | grep -v 'pot$'` ; do
  #   xgettext -j -o ./src/share/locale/en/LC_MESSAGES/OPNsense.pot "${i}"
  # done

For this commit, the old header was retained- it will be deleted in the next commit, for ease of scriptable maintinence of the canonical english .pot file.

Signed-off-by: Isaac (.ike) Levy <ike@blackskyresearch.net>