firmware: `better' fix for IPv6 loopback allow; #271
`block quick' creates a chain-reaction that requires `pass quick' for
loopback traffic, even though we have non-quick loopback pass rules
below.  If `quick' ought to be used, non-quick should be removed
completely, but that is beyond the scope of this fix.

For now, rather pass all IPv6 traffic on the looback device as opposed
to passing all IPv6 traffic with an IPv6 loopback address.