users: the infamous all group kicks the proverbial bucket
The all group was previously removed from the root account due
to the fact that it won't be applied to new user accounts anymore.
This isn't bad as far as we've seen so we went a bit further,
removing the group altogether.  Nice side effect is that a lot of
conditional code just disappears, so I doubt anyone will have any
hard feelings later.