captive portal: bye bye my love
The captive portal is being replaced!  This hooks up the
new page into the services section while ditching most of
the old code.  There'll be no migration, if you use package
`opnsense-devel' and the captive portal you'll have to switch
to `opnsense' or migrate to the new code.  Beware that the
new captive portal is going to be a huge step forward but
features will trickle in week after week until it is stable
enough to merge it.  Latest release date is going to be 16.1.