logs: tab conversion
Shuffled three of them back to status for the last round of
fire bombing before 15.7.17.  This is not going to be in the
stable version, but I want to have this ready for development
version feedback.  This is so much nicer without the tick
wrapping of tabs and pills and what not...