rc: allow to disable trim for the brave
Requires fiddling with /etc/fstab, at least that is a file
that we never touch.  It's harder for installs that need
TRIM to be off all the time as the first boot does actually
enable it.  That means after the install is complete it needs
to be CTRL+C'ed and the fstab entry modified before reboot.

PR: https://forum.opnsense.org/index.php?topic=3044