interfaces: remove pppoe reset cron handling... #911
... Ad already annotated a few spots, the whole thing is beyond
repair and I dread extending it to PPP and others which has been
asked for a few times now.  We will have to find another way.

To paraphrase Star Trek: He's dead, Jim.

The upside is that the cron config items are flushed by default
now, later we can remove this as we don't care what is in that
unused config section anyway.