ui: tip toeing around mobile issues for #1003
This tries to address table overflow problems we have and is
only meant as a base for further discussion.  It gets rid of
table overflow, e.g. in the dashboard on small mobile screens
by setting "table-layout: fixed;" for all tables.  This breaks
the stacked-to-horizontal emulation, so the workaround there
is to switch it off, which requires a proper viewport detection.

While there, fiddle with the logs template, but for these non-
hooked stacked-to-horizontal tables the fix for table-layout
can't be undone, because that would break the original fix in
the first place.

Moral of the story: tables are deprecated.  ;)