interfaces: defang netgraph attach/detach for #1188
We don't particularly need this code and the utility as netgraph
is always attached on load as long as we don't detach and the
attach per interface is a custom feature that requires a kernel
patch we don't want to keep around for 17.1.

Only thing we need to measure is the performance impact of not
detaching netgraph.  This only pertains to ethernet devices
anyway.

We do, however, keep interface_netgraph_needed() around for
reference as it's harmless by default.