rc: sshd rework... #1200
Uh, oh: the keys are not purged on factory reset.  Since we
now store the explicit host key file names in the config,
it's better to stop writing keys to /usr/local/etc/ssh and
just keep the backup where it was.  It's fully compatible
with our approach.

While there, remove the need to softcode commands in variables
as this was done to migrate from base OpenSSH to ports OpenSSH,
which was done in 2015.