system: finally split up use_mfs_tmpvar for good
This is an inplace config migration.  The mount still behaves
the same way and a config save from the page migrates the
options.  We're not in a rush here, but the reason is that
we should have clean options and the new 17.1 images for nano
will depend on the new style.