pkg: split @shadow and @sample syntax, slight differences
Can't guarantee this is the final state, but opnsense-bootstrap
stopping over lingering repository information without the finger-
prints in place is not very helpful either.

Now we have a working copy of the file we need, it is already
reconfigured after install and it now also reverts FreeBSD.conf
back to its required state--which is off--to prevent problems.