openvpn: fix config handling on PHP 7.1 #1733
The day has come that this "bootstrap" logic broke horribly,
preventing us from upgrading to PHP 7.1 with the initial 17.7.

Add a config_read_array() function that emulates *all* potentially
required steps and grabs the array reference so that the code can
be migrated easily.

Hopefully this also works on PHP 7.0.  ;)