gateways: ipv6 ll where appropriate
Normally, if we have an IPv6 link local monitor we use the link local
address which is always available.  If not a link local address, try
to use a global address.  If we don't have one, we can't reach the
outside anyway.

PR: https://forum.opnsense.org/index.php?topic=6028.0