intrusion detection: default syslog level seems to be Error, not Info
So do the sensible thing and use Notice which is what global logging
is currently using in the hopes to a align syslog with eve log.

Completely unsure if this helps though.  The docs don't describe this
and nobody else tinkering with suricata.yaml came up with a solution?

PR: https://forum.opnsense.org/index.php?topic=10140.0