gateways, reimplement into new class (https://github.com/opnsense/core/issues/2279)
Its work in progress, but the basic idea is to make sure we can easily output a list of all gateways (ipv4+ipv6) in the system, which is always correctly ordered by priority.
As soon as this works, we should be able to propogate ipprotocol to the gateway groups as well and easily iterate over the list to deterimine which gateways should be used at any time.

When settings are available in gateways->gateway_item, they should logically precede the ones generated dynamically.