interfaces: change the load order to be less error prone
Trackers shouldn't be special in GIF and GRE and bridges.  In
bridges there's also no use of VIPs so the code simplifies quite
a bit.  Load all exceptions in a row with the "unhandled" ones
still at the top of the function.

What needs to be seen is if VIPs latch on to interfaces that
are handled as exceptions correctly, or else we will have to
reload them alongside again.