Gateways: support interface targets (without addresses), but don't pinpoint it on specific drivers. for https://github.com/opnsense/core/pull/3983
Some tunnel interface types, such as Wireguard and Tinc do support sending traffic to the interface without an intermediate host. Since we don't want to add different static checks (and would like to get rid of the ones there for legacy reasons), it's probably better to add an option here.