MVC: add locking in JsonKeyValueStoreField type.
Although in theory the current stat() should at least make sure the same thread wouldn't execute the same action twice, it seems in reality actions are being executed
for every record in a set. Calling stat() after write+flush seems to return the previous status in stead of the one just written to disk (ufs issue on fbsd?). To prevent this from happening, use fstat() and lock the target
file while processing.

Found this with the firewall api plugin enabled and a set of rules in it, which seemed to trigger "list gateways" for every record.